+++ mqtt/mqtt-bridge/src/controller/mod.rs
@@ -1,32 +1,28 @@
-#![allow(unused_imports, unused_variables, dead_code)]
-use std::collections::HashMap;
+mod bridges;
+
+use bridges::Bridges;
 
 use futures_util::{
-    future::{self, BoxFuture, Either, Map},
-    pin_mut,
-    stream::{FuturesUnordered, Stream},
-    FusedStream, FutureExt, StreamExt,
+    future::{self, Either},
+    pin_mut, FusedStream, StreamExt,
 };
-use thiserror::Error;
-use tokio::{
-    sync::mpsc::{self, UnboundedReceiver, UnboundedSender},
-    task::{JoinError, JoinHandle},
-};
-use tracing::{debug, error, info, info_span, warn};
-use tracing_futures::Instrument;
+use tokio::sync::mpsc::{self, UnboundedReceiver, UnboundedSender};
+use tracing::{debug, error, info, warn};
 
 use crate::{
-    bridge::{Bridge, BridgeError, BridgeHandle},
-    config_update::BridgeUpdate,
-    config_update::{BridgeControllerUpdate, ConfigUpdater},
-    persist::StreamWakeableState,
+    bridge::{Bridge, BridgeError},
+    config_update::BridgeControllerUpdate,
     settings::BridgeSettings,
-    settings::ConnectionSettings,
 };
 
 const UPSTREAM: &str = "$upstream";
 
-/// Controller that handles the settings and monitors changes, spawns new Bridges and monitors shutdown signal.
+/// `BridgeController` controls lifetime of bridges: start/stop and update
+/// forwarding rules.
+///
+/// Controller handles monitors settings updates and starts a new `Bridge` or
+/// stops running `Bridge` if the number of bridges changes. In addition it
+/// prepares changes in forwarding rules and applies them to `Bridge` if require.
 pub struct BridgeController {
     handle: BridgeControllerHandle,
     messages: UnboundedReceiver<BridgeControllerMessage>,
@@ -71,7 +67,7 @@ impl BridgeController {
         let mut no_bridges = bridges.is_terminated();
 
         loop {
-            let any_bridge = if no_bridges {
+            let wait_bridge_or_pending = if no_bridges {
                 // if no active bridges available, wait only for a new messages arrival
                 Either::Left(future::pending())
             } else {
@@ -79,7 +75,7 @@ impl BridgeController {
                 Either::Right(bridges.next())
             };
 
-            match future::select(messages.select_next_some(), any_bridge).await {
+            match future::select(messages.select_next_some(), wait_bridge_or_pending).await {
                 Either::Left((BridgeControllerMessage::BridgeControllerUpdate(update), _)) => {
                     process_update(update, &mut bridges).await
                 }
@@ -92,7 +88,7 @@ impl BridgeController {
                     match bridge {
                         Ok(Ok(_)) => debug!("bridge {} exited", name),
                         Ok(Err(e)) => warn!(error = %e, "bridge {} exited with error", name),
-                        Err(e) => warn!(error = %e, "bridge {} paniced ", name),
+                        Err(e) => warn!(error = %e, "bridge {} panicked ", name),
                     }
 
                     info!("restarting bridge {}...", name);
@@ -108,7 +104,7 @@ impl BridgeController {
                     }
                 }
                 Either::Right((None, _)) => {
-                    // first time we resolve any_bridges future it returns None
+                    // first time we resolve bridge future it returns None
                     no_bridges = true;
                 }
             }
@@ -135,120 +131,16 @@ async fn process_update(update: BridgeControllerUpdate, bridges: &mut Bridges) {
     }
 }
 
-type BridgeFuture = BoxFuture<'static, (String, Result<Result<(), BridgeError>, JoinError>)>;
-
-#[derive(Default)]
-pub struct Bridges {
-    bridge_handles: HashMap<String, BridgeHandle>,
-    config_updaters: HashMap<String, ConfigUpdater>,
-    bridges: FuturesUnordered<BridgeFuture>,
-}
-
-impl Bridges {
-    async fn start_bridge<S>(&mut self, bridge: Bridge<S>, settings: &ConnectionSettings)
-    where
-        S: StreamWakeableState + Send + 'static,
-    {
-        let name = settings.name().to_owned();
-
-        let bridge_handle = bridge.handle();
-        let mut config_updater = ConfigUpdater::new(bridge.handle());
-
-        let bridge_name = settings.name().to_owned();
-        let upstream_bridge = bridge
-            .run()
-            .instrument(info_span!("bridge", name = %bridge_name));
-
-        // bridge running before sending initial settings
-        let task = tokio::spawn(upstream_bridge).map({
-            let name = name.clone();
-            move |res| (name, res)
-        });
-
-        // send initial subscription configuration
-        let update = BridgeUpdate::new(name.clone(), settings.subscriptions(), settings.forwards());
-        if let Err(e) = config_updater.send_update(update).await {
-            error!("failed to send initial subscriptions for {}. {}", name, e);
-        }
-
-        self.bridge_handles.insert(name.clone(), bridge_handle);
-        self.config_updaters.insert(name.clone(), config_updater);
-        self.bridges.push(Box::pin(task));
-    }
-
-    async fn send_update(&mut self, update: BridgeUpdate) {
-        if let Some(config) = self.config_updaters.get_mut(update.name()) {
-            if let Err(e) = config.send_update(update).await {
-                error!("error sending bridge update {:?}", e);
-            }
-        }
-    }
-
-    async fn shutdown_all(&mut self) {
-        debug!("sending shutdown request to all bridges...");
-
-        // sending shutdown signal to each bridge
-        let shutdowns = self
-            .bridge_handles
-            .drain()
-            .map(|(_, handle)| handle.shutdown());
-        future::join_all(shutdowns).await;
-
-        // wait until all bridges finish
-        while let Some((name, bridge)) = self.bridges.next().await {
-            match bridge {
-                Ok(Ok(_)) => debug!("bridge {} exited", name),
-                Ok(Err(e)) => warn!(error = %e, "bridge {} exited with error", name),
-                Err(e) => warn!(error = %e, "bridge {} panicked ", name),
-            }
-        }
-        debug!("all bridges exited");
-    }
-}
-
-impl Stream for Bridges {
-    type Item = (
-        String,
-        Result<Result<(), BridgeError>, tokio::task::JoinError>,
-    );
-
-    fn poll_next(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Option<Self::Item>> {
-        let poll = self.bridges.poll_next_unpin(cx);
-
-        // remove redundant handlers when bridge exits
-        if let std::task::Poll::Ready(Some((name, _))) = &poll {
-            self.bridge_handles.remove(name);
-            self.config_updaters.remove(name);
-        }
-
-        poll
-    }
-}
-
-impl FusedStream for Bridges {
-    fn is_terminated(&self) -> bool {
-        self.bridges.is_terminated()
-    }
-}
-
 impl Default for BridgeController {
     fn default() -> Self {
         Self::new()
     }
 }
 
+/// Provides a convenient handle to send control messages to `BridgeController`.
 #[derive(Clone, Debug)]
 pub struct BridgeControllerHandle {
     sender: UnboundedSender<BridgeControllerMessage>,
-}
-
-#[derive(Debug)]
-pub enum BridgeControllerMessage {
-    BridgeControllerUpdate(BridgeControllerUpdate),
-    Shutdown,
 }
 
 impl BridgeControllerHandle {
@@ -269,7 +161,15 @@ impl BridgeControllerHandle {
     }
 }
 
-#[derive(Debug, Error)]
+/// Control message for `BridgeController`.
+#[derive(Debug)]
+pub enum BridgeControllerMessage {
+    BridgeControllerUpdate(BridgeControllerUpdate),
+    Shutdown,
+}
+
+/// Error for `BridgeController`.
+#[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error("An error occurred sending a message to the controller.")]
     SendControllerMessage(#[source] tokio::sync::mpsc::error::SendError<BridgeControllerMessage>),
